Restricted Pages

+++ base/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.http import HttpResponse
@@ -68,6 +69,8 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     #working on submitting data
@@ -81,9 +84,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 #updating the room
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed Here!!')
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -94,8 +101,14 @@ def updateRoom(request, pk):
     context = {'form': form, 'room': room}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed Here!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
